Minor cleanups * Corrected a few potential crash points. * Corrected a possible cast of overlapping instances IDs. * Changed instance database ID variable name to not get confused with live instance data ID variable name. * Added a warning during parsing about exceeding maximum amount of instances. * Corrected a client instance ID lookup. * Corrected instance map naming. Thanks to @Atemo's review!

+++ src/map/clif.hpp
@@ -822,10 +822,10 @@ void clif_sendbgemblem_area(struct map_session_data *sd);
 void clif_sendbgemblem_single(int fd, struct map_session_data *sd);
 
 // Instancing
-void clif_instance_create(unsigned short instance_id, int num);
-void clif_instance_changewait(unsigned short instance_id, int num);
-void clif_instance_status(unsigned short instance_id, unsigned int limit1, unsigned int limit2);
-void clif_instance_changestatus(unsigned int instance_id, int type, unsigned int limit);
+void clif_instance_create(int instance_id, int num);
+void clif_instance_changewait(int instance_id, int num);
+void clif_instance_status(int instance_id, unsigned int limit1, unsigned int limit2);
+void clif_instance_changestatus(int instance_id, int type, unsigned int limit);
 
 // Custom Fonts
 void clif_font(struct map_session_data *sd);
